+++ src/test/java/toyproject/personal/englishconversation/service/ChatGPTServiceTest.java
@@ -4,6 +4,7 @@ import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.context.annotation.Import;
 import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -14,6 +15,7 @@ import toyproject.personal.englishconversation.config.chatgpt.ChatGPTContent;
 import toyproject.personal.englishconversation.controller.dto.chatgpt.ChatGPTRequestDto;
 import toyproject.personal.englishconversation.controller.dto.chatgpt.ChatGPTResponseDto;
 import toyproject.personal.englishconversation.controller.dto.chatgpt.Message;
+import toyproject.personal.englishconversation.NoSecurityTestConfig;
 import toyproject.personal.englishconversation.domain.message.GPTMessage;
 
 import java.util.HashMap;
@@ -26,6 +28,7 @@ import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.*;
 
 @SpringBootTest
+@Import(NoSecurityTestConfig.class)
 class ChatGPTServiceTest {
 
     @Mock
