+++ src/test/java/toyproject/personal/englishconversation/service/ChatServiceTest.java
@@ -7,6 +7,8 @@ import org.springframework.boot.test.context.SpringBootTest;
 import toyproject.personal.englishconversation.config.chatgpt.ChatGPTContent;
 import toyproject.personal.englishconversation.controller.dto.chatgpt.ChatGPTRequestDto;
 import toyproject.personal.englishconversation.controller.dto.chatgpt.Message;
+import org.springframework.context.annotation.Import;
+import toyproject.personal.englishconversation.NoSecurityTestConfig;
 import toyproject.personal.englishconversation.domain.message.GPTMessage;
 
 import java.io.IOException;
@@ -20,6 +22,7 @@ import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.*;
 
 @SpringBootTest
+@Import(NoSecurityTestConfig.class)
 class ChatServiceTest {
 
     @Mock
